refactor: move `emit` with `windowLabel` back to helpers
